+++ packages/cli-plugin-lint/src/add-prettier-config.ts
@@ -3,129 +3,75 @@
  * @description add prettier config in .prettier.js
  */
 
-import { runPrompts } from "./prompts";
-import fs from "fs-extra";
-import ora from "ora";
-import { resolve } from "path";
-import { prettierDeps } from "./config";
-import chalk from "chalk";
-import {
-  getFilenameForDirectory,
-  loadConfigFile,
-  write,
-} from "./utils/config-file";
-import { Logger } from "./logger";
-import { installSaveDev } from "./utils/npm-utils";
+import fs from 'fs-extra';
+import ora from 'ora';
+import { resolve } from 'path';
+import chalk from 'chalk';
+import { prettierDeps } from './config';
+import { getFilenameForDirectory, loadConfigFile, write } from './utils/config-file';
+import { Logger } from './logger';
+import { installSaveDev } from './utils/npm-utils';
 
-const installOraInstance = ora("install-prettier");
+const installOraInstance = ora('install-prettier');
+const PRETTIER_CONFIG_PATH = resolve(__dirname, '../config/.prettierrc.js');
 
-const mainOptions = [
-  {
-    type: "select",
-    name: "run",
-    message: "How should Prettier be run?",
-    choices: [
-      {
-        message: "By the Prettier CLI/plugin",
-        name: "prettier",
-      },
-      {
-        message: "By ESLint",
-        name: "eslint",
-      },
-    ],
-  },
-];
-
-const PRETTIER_CONFIG_PATH = resolve(__dirname, "../config/.prettierrc.js");
-
-const addPrettierToExtends = (
-  eslintObj: { extends: any },
-  prettierExtensionStr: string
-) => {
+const addPrettierToExtends = (eslintObj: { extends: any }, prettierExtensionStr: string) => {
   if (eslintObj && eslintObj.extends && Array.isArray(eslintObj.extends)) {
     return [...eslintObj.extends, prettierExtensionStr];
   } else {
-    return eslintObj.extends
-      ? [eslintObj.extends, prettierExtensionStr]
-      : prettierExtensionStr;
+    return eslintObj.extends ? [eslintObj.extends, prettierExtensionStr] : prettierExtensionStr;
   }
 };
-
-const addPrettierToRules = (
-  prettierObj: any,
-  eslintObj: { rules: any; extends: any }
-) => ({
-  ...eslintObj.rules,
-  "prettier/prettier": prettierObj ? ["error", prettierObj] : "error",
-});
 
 const extendESLintConfig = (
   eslintObj: { extends: any; rules: any } = { extends: [], rules: {} }
 ) => ({
   ...eslintObj,
-  extends: addPrettierToExtends(eslintObj, "prettier"),
+  extends: addPrettierToExtends(eslintObj, 'prettier')
 });
 
-const prettierQuestions = () =>
-  runPrompts<{ run: string }>(mainOptions)
-    .then(({ run }) => ({ run }))
-    .catch((error) => console.error(error));
-
 const createPrettierrc = (prettier: string, dir: string = process.cwd()) => {
-  const prettierDir = resolve(dir, "./.prettierrc.js");
+  const prettierDir = resolve(dir, './.prettierrc.js');
   return fs
     .ensureFile(prettierDir)
-    .then(() => {
-      return fs
+    .then(() =>
+      fs
         .outputFile(prettierDir, prettier)
         .then(() =>
-          Logger.info(
-            chalk.yellow(
-              `\n👏 Created .prettierrc.js, please check for sure \n`
-            )
-          )
+          Logger.info(chalk.yellow(`👏 Created .prettierrc.js, please check for sure \n`))
         )
         .catch((error) => {
           console.log(error);
-        });
-    })
-    .catch((err: any) => console.log(err));
+        })
+    )
+    .catch((err: unknown) => console.log(err));
 };
 
 /**
  * add prettier config
  */
 
-const configPrettierRC = async (pmTool?: string) => {
+const configPrettierRC = async (pmTool?: string): Promise<void> => {
   try {
-    for (const dep in prettierDeps) {
-      installOraInstance.start("安装 prettier 依赖\n" + dep + "@" + prettierDeps[dep]);
+    const preDenps = Object.keys(prettierDeps);
+    for (let index = 0; index < preDenps.length; index++) {
+      const dep = preDenps[index];
+      installOraInstance.start(`安装 prettier\n${dep}@${prettierDeps[dep]}`);
       await installSaveDev(dep, prettierDeps[dep], pmTool);
       installOraInstance.clear();
     }
+    installOraInstance.succeed('安装 prettier');
 
-    installOraInstance.succeed("安装 prettier 依赖");
-
-    // const answer = await prettierQuestions();
-    // if (answer) {
-    // const { run } = answer;
-    const run = "prettier"; // run prettier  By the Prettier CLI/plugin  default
-    // Logger.debug(chalk.green("Adding prettier config."));
     const file = getFilenameForDirectory(process.cwd());
     const eslint = loadConfigFile({ filePath: file }) || {
       extends: [],
-      rules: {},
+      rules: {}
     };
-
-    write(extendESLintConfig(eslint), file);
-
-    if (run === "prettier") {
-      const prettierConfig = await fs.readFile(PRETTIER_CONFIG_PATH, {
-        encoding: "utf-8",
-      });
-      await createPrettierrc(prettierConfig);
-    }
+    if (file) write(extendESLintConfig(eslint), file);
+    const prettierConfig = await fs.readFile(PRETTIER_CONFIG_PATH, {
+      encoding: 'utf-8'
+    });
+    await createPrettierrc(prettierConfig);
   } catch (error) {
     Logger.error(chalk.red(`${error}`));
     process.exit(0);