+++ packages/cli-plugin-lint/src/ci.ts
@@ -78,12 +78,13 @@ async function initHusky(pmTool: string) {
   const PRE_COMMIT_PATH = `${HUSKY_CONFIG_PATH}/pre-commit`;
 
   // install huksy
-  Object.keys(huskyCiDeps).forEach(async (dep) => {
-    installOraInstance.start(`安装 husky 依赖\n${dep}@${huskyCiDeps[dep]}`);
+  const huskyDenps = Object.keys(huskyCiDeps);
+  for (let index = 0; index < huskyDenps.length; index++) {
+    const dep = huskyDenps[index];
+    installOraInstance.start(`安装 husky\n${dep}@${huskyCiDeps[dep]}`);
     await installSaveDev(dep, huskyCiDeps[dep], pmTool);
     installOraInstance.clear();
-  });
-
+  }
   installOraInstance.succeed('安装 husky');
 
   fs.rmdirSync(HUSKY_CONFIG_PATH, { recursive: true });
@@ -99,7 +100,7 @@ async function initHusky(pmTool: string) {
     encoding: 'utf-8'
   });
 
-  if (spawnHinstall.stderr) {
+  if (spawnHinstall.status !== 0) {
     throw new Error(`husky 初始化失败 \n ${spawnHinstall.stderr}`);
   }
 
@@ -112,20 +113,23 @@ async function initHusky(pmTool: string) {
       encoding: 'utf-8'
     }
   );
-  if (spawnHadd.stderr) {
+  if (spawnHadd.status !== 0) {
     throw new Error(`husky add pre-commit hook failed \n ${spawnHadd.stderr}`);
   }
 
-  Logger.info(chalk.yellow(`\n👏 husky 配置完成, please check for sure. \n`));
+  Logger.info(chalk.yellow(`👏 husky 配置完成, please check for sure. \n`));
 }
 
 async function initLintstaged(projectType: string, supportTypeScript: boolean, pmTool: string) {
   // install huksy
-  Object.keys(lintstagedCiDeps).forEach(async (dep) => {
-    installOraInstance.start(`${dep}@${lintstagedCiDeps[dep]}`);
+  const lintstagedDenps = Object.keys(lintstagedCiDeps);
+  for (let index = 0; index < lintstagedDenps.length; index++) {
+    const dep = lintstagedDenps[index];
+    installOraInstance.start(`安装 lint-staged\n${dep}@${lintstagedCiDeps[dep]}`);
     await installSaveDev(dep, lintstagedCiDeps[dep], pmTool);
     installOraInstance.clear();
-  });
+  }
+  installOraInstance.succeed('安装 lint-staged');
 
   //init lint-staged config
   const suffix = ['js'];
@@ -151,10 +155,8 @@ async function initLintstaged(projectType: string, supportTypeScript: boolean, p
 
   // 需要删除旧的lint-staged配置
   mdifyConfigPackage();
-
   Logger.info(chalk.green(`当前lint文件为"${LINT_STAGED_CONFIG_NAME}",可根据项目具体情况调整`));
-
-  Logger.info(chalk.yellow(`\n👏 lint-staged 配置完成, please check for sure. \n`));
+  Logger.info(chalk.yellow(`👏 lint-staged 配置完成, please check for sure. `));
 }
 
 /**