+++ packages/cli-plugin-lint/CHANGELOG.md
@@ -3,6 +3,141 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.0.12 (2021-06-28)
+
+
+### Bug Fixes
+
+* fix sentry registry ([b33e5ef](https://github.com/cool-fe/winex-cli/commit/b33e5ef75f388f8883b912897c67ce28d0794a75))
+
+
+
+
+
+## 1.0.11 (2021-06-24)
+
+
+### Features
+
+* add log sentry ([4d8d56f](https://github.com/cool-fe/winex-cli/commit/4d8d56fd1e79b6e8dade1e8d61e25801fc255c86))
+
+
+
+
+
+## 1.0.10 (2021-06-17)
+
+
+### Bug Fixes
+
+* fix cli eslint ([4089c4f](https://github.com/cool-fe/winex-cli/commit/4089c4f01228572a8e019f897a1d90b156908eea))
+
+
+
+
+
+## [1.0.9](https://github.com/cool-fe/winex-cli/compare/v1.0.8...v1.0.9) (2021-06-11)
+
+
+### Bug Fixes
+
+* installDeps import bug ([e4a88ba](https://github.com/cool-fe/winex-cli/commit/e4a88bab1d38c84e683430d11653b0598d3f9545))
+
+
+
+
+
+## [1.0.8](https://github.com/cool-fe/winex-cli/compare/v1.0.7...v1.0.8) (2021-06-11)
+
+**Note:** Version bump only for package @winfe/cli-plugin-lint
+
+
+
+
+
+
+## [1.0.7](https://github.com/cool-fe/winex-cli/compare/v1.0.6...v1.0.7) (2021-06-11)
+
+
+### Bug Fixes
+
+* 优化日志样式和处理eslint error ([1d5082a](https://github.com/cool-fe/winex-cli/commit/1d5082a2d3cb95c7a0b7698698f9301b2405b275))
+
+
+
+
+
+## [1.0.6](https://github.com/cool-fe/winex-cli/compare/v1.0.4...v1.0.6) (2021-06-11)
+
+
+### Bug Fixes
+
+* fix eslint error ([d336e81](https://github.com/cool-fe/winex-cli/commit/d336e810e34a48810ea038d4632526312f067713))
+
+
+
+
+
+## [1.0.5](https://github.com/cool-fe/winex-cli/compare/v1.0.4...v1.0.5) (2021-06-11)
+
+
+### Bug Fixes
+
+* fix eslint error ([d336e81](https://github.com/cool-fe/winex-cli/commit/d336e810e34a48810ea038d4632526312f067713))
+
+
+
+
+
+## [1.0.4](https://github.com/cool-fe/winex-cli/compare/v1.0.3...v1.0.4) (2021-06-11)
+
+
+### Bug Fixes
+
+* perform text ([bc63d04](https://github.com/cool-fe/winex-cli/commit/bc63d04629a2b39a099e9ba2c69512869c072438))
+
+
+
+
+
+
+## 1.0.3 (2021-06-11)
+
+**Note:** Version bump only for package @winfe/cli-plugin-lint
+
+
+
+
+
+## 1.0.2 (2021-06-11)
+
+**Note:** Version bump only for package @winfe/cli-plugin-lint
+
+
+
+
+
+## [1.0.1](https://github.com/cool-fe/winex-cli/compare/v1.0.0...v1.0.1) (2021-06-11)
+
+
+### Bug Fixes
+
+* fix process.stderr ([0c655b2](https://github.com/cool-fe/winex-cli/commit/0c655b2f85efd5ae3535b5aa6119af65a9fa6eb0))
+* perform log style ([63b1411](https://github.com/cool-fe/winex-cli/commit/63b1411d9d712cfa222b616f13a42af5690f8e7c))
+
+
+
+
+
+
+# 1.0.0 (2021-06-10)
+
+**Note:** Version bump only for package @winfe/cli-plugin-lint
+
+
+
+
+
 # 1.0.0-beta.11 (2021-06-07)
 
 **Note:** Version bump only for package @winfe/cli-plugin-lint